Only add job if it does not exists
If using the jdbcstore and you use persistent jobdatamap, it will be
null the next system restart, because jobdetail is added with replace =
true. Now the system checks if the job already exists.
Furthermore a loadAfter is added for databaseMigration, when you want to
add the databaseschema of quartz using dbm and quartz is loaded before
dbm, it will fail. Loadafter makes sure this does not happen.